+++ nix/versions.nix
@@ -1,6 +1,6 @@
 {
-  fennel-stable = "1.4.1";
-  fennel-unstable = "1.4.2-dev";
+  fennel-stable = "1.4.2";
+  fennel-unstable = "1.5.0-dev";
   faith-stable = "0.1.2";
   faith-unstable = "0.1.3-dev";
   fnlfmt-stable = "0.3.1";